requirements.txt
click==8.0.3
python-dotenv==0.19.2
requests==2.26.0
PyGithub==1.55
openai==0.27.0
anthropic==0.3.0
rich==10.12.0
pydantic==1.10.9